Add files via upload
This contains the code for first version of chord extractor program. This program can identify chords played in a soundtrack of about 0.3 - 0.5 seconds.  Use the trainer to train for your database (change mypath to path where the database is present). The program is already trained though with 96.65% accuracy for the training set.The values of theta corresponding to the training is unrolled in theta.txt .